create options class
